% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpac.R
\name{clean_spaces}
\alias{clean_spaces}
\title{clean_spaces}
\usage{
clean_spaces(str)
}
\arguments{
\item{str}{an input string containing spaces that we want to replace with underscores.}
}
\value{
a string with no spaces.
}
\description{
clean_spaces
}
\examples{
clean_spaces("Thamer Aldawood")
}
