% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_corr.R
\name{compute_corr}
\alias{compute_corr}
\title{Title}
\usage{
compute_corr(data, var1, var2)
}
\arguments{
\item{data}{input dataset}

\item{var1}{variable 1}

\item{var2}{bariable 2}
}
\value{
the tidy correlation
}
\description{
Title
}
\examples{
compute_corr(data = faithful, var1 = eruptions, var2 = waiting)

}
